Transformation to CMake project. Gui project fixes. #396

+++ client/gui/backend/persistentstate.cpp
@@ -33,7 +33,7 @@ void PersistentState::save()
 {
     QSettings settings;
 
-    int size = state_.ByteSizeLong();
+    int size = (int)state_.ByteSizeLong();
     QByteArray arr(size, Qt::Uninitialized);
     state_.SerializeToArray(arr.data(), size);
